Split BusinessDetails into smaller components
