+++ src/Components/UI/Business/BusinessKnownFor.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+export default () => {
+  return (
+    <>
+      <h3 className="text-uppercase text-muted h6 mt-4 mb-3">Known for</h3>
+      <div>
+        <p>Quality Products</p>
+        <p>Fast Delivery</p>
+        <p>Sweet Packaging</p>
+      </div>
+    </>
+  );
+};
