+++ src/Components/UI/Business/RelatedBusinessList.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import RelatedBusinesses from "../../RelatedBusinesses";
+
+export default () => (
+  <>
+    <h3 className="text-uppercase text-muted h6 my-3">Related Businesses</h3>
+    <RelatedBusinesses />
+    <RelatedBusinesses />
+    <RelatedBusinesses />
+    <RelatedBusinesses />
+    <RelatedBusinesses />
+  </>
+);
