+++ src/Components/UI/Business/index.js
@@ -1,0 +1,11 @@
+import BusinessKnownFor from "./BusinessKnownFor";
+import BusinessRating from "./BusinessRating";
+import BusinessWebsite from "./BusinessWebsite";
+import RelatedBusinessList from "./RelatedBusinessList";
+
+export {
+  BusinessRating,
+  BusinessKnownFor,
+  BusinessWebsite,
+  RelatedBusinessList,
+};
